refactor(proxy/http): move `is_bad_request()` into `upgrade`
this is only used in once place, so as a brief chore before we move the
upgrade submodule out into its own crate, we pull `is_bad_request()`
next to its call site.
